python/requirements.txt
flask
argparse
psycopg2-binary
requests
python-keycloak
